圆点导航点优化超长显示需求

+++ frameworks/core/components_ng/pattern/swiper_indicator/dot_indicator/dot_indicator_modifier.h
@@ -26,7 +26,7 @@
 #include "core/components_ng/render/paint_wrapper.h"
 namespace OHOS::Ace::NG {
 constexpr int32_t ITEM_SIZE = 4;
-constexpr int32_t INIT_SIZE_RATE = 1.0f;
+constexpr float INIT_SIZE_RATE = 1.0f;
 
 enum class TouchBottomType {
     NONE = 0,
